Allow maketags to tag languages other than C/C++
Modify the ctags call in maketags to just use -R rather than the file list if
nothing in the list is being filtered.  If it's being filtered, continue to
use the file list, so that the filters are applied to ctags as well as cscope.
This can likely be improved.