Add test scripts for clean emerging
gnome-test:

This is a script that takes an atom, untars a clean chroot package, chroots into
it, and emerges that package.  It spits it's output to stdout/stderr, and
returns exit status based on success or failure.  It depends on dchroot (and
specifically on a new enough one to have the run command).  The chroot package
needs to be set up ahead of time.

gnome-test-all:

This script walks every entry in a file named pkglist, and runs gnome-test on
each atom in the file.  It tees it's output for each atom to a file in
~/gnome-test/current, and if it fails, stores that output in
~/gnome-test/failed/<atom-name>.  It does not quit on failure, but continues
with the next atom in the file.