Exactly match chroot name for stop
If you had two screen sessions, one of which was a superset of the name
of the chroot, then the grep on `screen -ls` would return them both,
causing "dchroot stop" to fail.  Fix that by matching the name exactly
by prefixing the patten with literal '.' and suffexing it with
whitespace.